Add base class for Siganet
